+++ apps/web/src/App.vue
@@ -3,6 +3,7 @@ import {
 	darkTheme,
 	NConfigProvider,
 	NMessageProvider,
+	NModalProvider,
 	NNotificationProvider,
 } from "naive-ui";
 import BHeader from "./components/BHeader.vue";
@@ -16,35 +17,37 @@ import Rainbow from "./components/rainbow.vue";
       :theme="darkTheme"
       preflight-style-disabled
   >
-    <NMessageProvider>
-      <NNotificationProvider :max="3">
-        <div class="fixed">
-          <Rainbow />
-        </div>
-
-        <BMenu>
-          <div class="app pb-10 xl:pb-0">
-            <BHeader />
-
-            <RouterView v-slot="{ Component }">
-              <template v-if="Component">
-                <Transition mode="out-in">
-                  <KeepAlive>
-                    <Suspense>
-                      <component :is="Component"></component>
-
-                      <template #fallback>
-                        Loading...
-                      </template>
-                    </Suspense>
-                  </KeepAlive>
-                </Transition>
-              </template>
-            </RouterView>
+    <NModalProvider>
+      <NMessageProvider>
+        <NNotificationProvider :max="3">
+          <div class="fixed">
+            <Rainbow />
           </div>
-        </BMenu>
-      </NNotificationProvider>
-    </NMessageProvider>
+
+          <BMenu>
+            <div class="app pb-10 xl:pb-0">
+              <BHeader />
+
+              <RouterView v-slot="{ Component }">
+                <template v-if="Component">
+                  <Transition mode="out-in">
+                    <KeepAlive>
+                      <Suspense>
+                        <component :is="Component"></component>
+
+                        <template #fallback>
+                          Loading...
+                        </template>
+                      </Suspense>
+                    </KeepAlive>
+                  </Transition>
+                </template>
+              </RouterView>
+            </div>
+          </BMenu>
+        </NNotificationProvider>
+      </NMessageProvider>
+    </NModalProvider>
   </NConfigProvider>
 </template>